@@ -1,7 +1,11 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Box, Typography, Button, Card, CardContent, Avatar } from '@mui/material';
 import { makeStyles } from '@material-ui/core';
 import LocationOnIcon from '@mui/icons-material/LocationOn';
+import Swal from 'sweetalert2';
+import utils from '../utils';
+import { useDispatch } from 'react-redux';
+import { actions } from '../redux/actions';
 
 const useStyles = makeStyles({
     locationCard: {
@@ -27,8 +31,11 @@ const useStyles = makeStyles({
     continueButton: {
         borderRadius: '25px !important',
         textTransform: 'none !important',
-        backgroundColor: '#4CAF50 !important',
+        backgroundColor: (props) => (props.isButtonDisabled ? 'gray' : '#4CAF50 !important'),
         color: 'white !important',
+        '&:hover': {
+            backgroundColor: (props) => (props.isButtonDisabled ? 'gray' : '#4CAF50 !important'),
+        },
     },
     cardsContainer: {
         overflowY: 'auto !important',
@@ -42,44 +49,77 @@ const useStyles = makeStyles({
         margin: '10px 0 !important',
         boxShadow: '#ffffff !important',
         borderRadius: '20px !important',
+        cursor: 'pointer',
+        transition: 'background-color 0.3s, color 0.3s'
     },
     cardContent: {
-        display: 'flex !important',
-        alignItems: 'center !important',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
     },
     avatar: {
         marginRight: '16px !important',
     },
     distanceText: {
         fontWeight: 'bold !important',
-        padding: "10px !important"
+        padding: "10px !important",
+        marginLeft: 2
     },
     locationText: {
         color: "#252B5C !important",
         fontSize: "23px  !important"
+    },
+    selectedCard: {
+        backgroundColor: '#4CAF50 !important',
+        color: 'white !important',
     }
 });
 
-export const LocationLayout = () => {
-    const classes = useStyles();
-    const distances = [
-        '0 - 50m',
-        '50 - 100m',
-        '100 - 150m',
-        '150 - 200m'
-    ];
+export const LocationLayout = ({ onConfirm }) => {
+    const dispatch = useDispatch();
+    const [selectedDistance, setSelectedDistance] = useState(null);
+    const [isButtonDisabled, setIsButtonDisabled] = useState(true);
+    const classes = useStyles({ isButtonDisabled });
+
+    const handleCardClick = (distance) => {
+        setSelectedDistance(distance);
+        setIsButtonDisabled(false);
+    };
+
+    const handleContinueClick = () => {
+        Swal.fire('Security guard notified!', 'The security guard has been notified to open the gate.', 'success').then(() => {
+            notifyAdmin(selectedDistance);
+            dispatch({ type: actions.CLOSE_LOCATION_CARDS, payload: false });
+            onConfirm();
+        });
+    };
+
+    const notifyAdmin = (distance) => {
+        // Logic to send alert to the admin and security app
+        console.log(`Admin notified to open the gate. Distance: ${distance}`);
+        // You might want to implement this with a real-time service like Firebase or a server API.
+    };
 
     return (
         <Box className={classes.locationCard}>
             <Box className={classes.distanceSelectHeader}>
                 <Typography variant="h6" className={classes.locationText}>{"Select Location Distance"}</Typography>
-                <Button variant="contained" color="primary" className={classes.continueButton}>
-                    {"  Continue"}
+                <Button
+                    variant="contained"
+                    className={classes.continueButton}
+                    onClick={handleContinueClick}
+                    disabled={isButtonDisabled}
+                >
+                    {"Continue"}
                 </Button>
             </Box>
             <Box className={classes.cardsContainer}>
-                {distances.map((distance, index) => (
-                    <Card key={index} className={classes.card}>
+                {utils.distances.map((distance, index) => (
+                    <Card
+                        key={index}
+                        className={`${classes.card} ${selectedDistance === distance ? classes.selectedCard : ''}`}
+                        onClick={() => handleCardClick(distance)}
+                    >
                         <CardContent className={classes.cardContent}>
                             <Avatar>
                                 <LocationOnIcon />
